refactor : 입력 값 검증 로직 분리
기존의 유효성 검증 로직을 별도의 파일로 분리
