+++ docs/README.md
@@ -16,7 +16,7 @@
 - [ ] 테스트 코드 구현
 - [ ] 에러 처리
   - [ ] "[ERROR]" 이 들어가는 예외 throw
-- [ ] 리팩토링
+- [x] 리팩토링
 
 ### 🚀 기능 요구 사항
